consegui fazer com mais componentsgit add .
